feat: add contract is now concat contract which concatenates two strings
