--- research_share/src/contracts/add_bytecode.json
@@ -1,6 +1,0 @@
-{
-	"linkReferences": {},
-	"object": "608060405234801561001057600080fd5b506040516101b53803806101b58339818101604052604081101561003357600080fd5b810190808051906020019092919080519060200190929190505050816001819055508060028190555050506101488061006d6000396000f3fe608060405234801561001057600080fd5b506004361061004c5760003560e01c80632ee7f53614610051578063653721471461007f5780637cac97361461009d578063eb4a77d3146100a7575b600080fd5b61007d6004803603602081101561006757600080fd5b81019080803590602001909291905050506100d5565b005b6100876100df565b6040518082815260200191505060405180910390f35b6100a56100e5565b005b6100d3600480360360208110156100bd57600080fd5b81019080803590602001909291905050506100fb565b005b8060018190555050565b60005481565b6100f3600154600254610105565b600081905550565b8060028190555050565b600081830190509291505056fea2646970667358221220dbeb97bcca9cc704cfd25e9adec76159a976d5416db9352cfc397c86c28e51c364736f6c63430007000033",
-	"opcodes": "PUSH1 0x80 PUSH1 0x40 MSTORE CALLVALUE DUP1 ISZERO PUSH2 0x10 JUMPI PUSH1 0x0 DUP1 REVERT JUMPDEST POP PUSH1 0x40 MLOAD PUSH2 0x1B5 CODESIZE SUB DUP1 PUSH2 0x1B5 DUP4 CODECOPY DUP2 DUP2 ADD PUSH1 0x40 MSTORE PUSH1 0x40 DUP2 LT ISZERO PUSH2 0x33 JUMPI PUSH1 0x0 DUP1 REVERT JUMPDEST DUP2 ADD SWAP1 DUP1 DUP1 MLOAD SWAP1 PUSH1 0x20 ADD SWAP1 SWAP3 SWAP2 SWAP1 DUP1 MLOAD SWAP1 PUSH1 0x20 ADD SWAP1 SWAP3 SWAP2 SWAP1 POP POP POP DUP2 PUSH1 0x1 DUP2 SWAP1 SSTORE POP DUP1 PUSH1 0x2 DUP2 SWAP1 SSTORE POP POP POP PUSH2 0x148 DUP1 PUSH2 0x6D PUSH1 0x0 CODECOPY PUSH1 0x0 RETURN INVALID PUSH1 0x80 PUSH1 0x40 MSTORE CALLVALUE DUP1 ISZERO PUSH2 0x10 JUMPI PUSH1 0x0 DUP1 REVERT JUMPDEST POP PUSH1 0x4 CALLDATASIZE LT PUSH2 0x4C JUMPI PUSH1 0x0 CALLDATALOAD PUSH1 0xE0 SHR DUP1 PUSH4 0x2EE7F536 EQ PUSH2 0x51 JUMPI DUP1 PUSH4 0x65372147 EQ PUSH2 0x7F JUMPI DUP1 PUSH4 0x7CAC9736 EQ PUSH2 0x9D JUMPI DUP1 PUSH4 0xEB4A77D3 EQ PUSH2 0xA7 JUMPI JUMPDEST PUSH1 0x0 DUP1 REVERT JUMPDEST PUSH2 0x7D PUSH1 0x4 DUP1 CALLDATASIZE SUB PUSH1 0x20 DUP2 LT ISZERO PUSH2 0x67 JUMPI PUSH1 0x0 DUP1 REVERT JUMPDEST DUP2 ADD SWAP1 DUP1 DUP1 CALLDATALOAD SWAP1 PUSH1 0x20 ADD SWAP1 SWAP3 SWAP2 SWAP1 POP POP POP PUSH2 0xD5 JUMP JUMPDEST STOP JUMPDEST PUSH2 0x87 PUSH2 0xDF JUMP JUMPDEST PUSH1 0x40 MLOAD DUP1 DUP3 DUP2 MSTORE PUSH1 0x20 ADD SWAP2 POP POP PUSH1 0x40 MLOAD DUP1 SWAP2 SUB SWAP1 RETURN JUMPDEST PUSH2 0xA5 PUSH2 0xE5 JUMP JUMPDEST STOP JUMPDEST PUSH2 0xD3 PUSH1 0x4 DUP1 CALLDATASIZE SUB PUSH1 0x20 DUP2 LT ISZERO PUSH2 0xBD JUMPI PUSH1 0x0 DUP1 REVERT JUMPDEST DUP2 ADD SWAP1 DUP1 DUP1 CALLDATALOAD SWAP1 PUSH1 0x20 ADD SWAP1 SWAP3 SWAP2 SWAP1 POP POP POP PUSH2 0xFB JUMP JUMPDEST STOP JUMPDEST DUP1 PUSH1 0x1 DUP2 SWAP1 SSTORE POP POP JUMP JUMPDEST PUSH1 0x0 SLOAD DUP2 JUMP JUMPDEST PUSH2 0xF3 PUSH1 0x1 SLOAD PUSH1 0x2 SLOAD PUSH2 0x105 JUMP JUMPDEST PUSH1 0x0 DUP2 SWAP1 SSTORE POP JUMP JUMPDEST DUP1 PUSH1 0x2 DUP2 SWAP1 SSTORE POP POP JUMP JUMPDEST PUSH1 0x0 DUP2 DUP4 ADD SWAP1 POP SWAP3 SWAP2 POP POP JUMP INVALID LOG2 PUSH5 0x6970667358 0x22 SLT KECCAK256 0xDB 0xEB SWAP8 0xBC 0xCA SWAP13 0xC7 DIV 0xCF 0xD2 0x5E SWAP11 0xDE 0xC7 PUSH2 0x59A9 PUSH23 0xD5416DB9352CFC397C86C28E51C364736F6C6343000700 STOP CALLER ",
-	"sourceMap": "64:506:0:-:0;;;143:85;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;197:4;191:3;:10;;;;217:4;211:3;:10;;;;143:85;;64:506;;;;;;"
-}
